Feature1

+++ contracts/eventTransactionRecord.sol
@@ -7,12 +7,12 @@ contract TransactionContract {
         string transactionType;
         string eventId;
         address Recorder;
-        mapping(string => uint256) params;
+        mapping(string => int256) params;
     }
     
     Transaction[] public transactions;
 
-    function E00010001(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _EP003, uint256 _EP005) public {
+    function E00010001(string memory _eventId, int256 _EP001, int256 _EP002, int256 _EP003, int256 _EP005) public {
         uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -23,11 +23,11 @@ contract TransactionContract {
         transaction.params["EP001"] = _EP001;
         transaction.params["EP002"] = _EP002;
         transaction.params["EP003"] = _EP003;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
         transaction.params["EP005"] = _EP005;
     }
 
-    function E00010002(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _EP003, uint256 _EP005) public {
+    function E00010002(string memory _eventId, int256 _EP001, int256 _EP002, int256 _EP003, int256 _EP005) public {
         uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -38,11 +38,11 @@ contract TransactionContract {
         transaction.params["EP001"] = _EP001;
         transaction.params["EP002"] = _EP002;
         transaction.params["EP003"] = _EP003;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
         transaction.params["EP005"] = _EP005;
     }
 
-    function E00010003(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _EP003, uint256 _EP005) public {
+    function E00010003(string memory _eventId, int256 _EP001, int256 _EP002, int256 _EP003, int256 _EP005) public {
         uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -53,11 +53,11 @@ contract TransactionContract {
         transaction.params["EP001"] = _EP001;
         transaction.params["EP002"] = _EP002;
         transaction.params["EP003"] = _EP003;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
         transaction.params["EP005"] = _EP005;
     }
 
-    function E00010004(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _EP003, uint256 _EP004, uint256 _EP006, uint256 _EP007) public {
+    function E00010004(string memory _eventId, int256 _EP001, int256 _EP002, int256 _EP003, int256 _EP004, int256 _EP006, int256 _EP007) public {
         uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -69,12 +69,12 @@ contract TransactionContract {
         transaction.params["EP002"] = _EP002;
         transaction.params["EP003"] = _EP003;
         transaction.params["EP004"] = _EP004;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
         transaction.params["EP006"] = _EP006;
         transaction.params["EP007"] = _EP007;
     }
 
-    function E00010005(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _EP003, uint256 _EP004, uint256 _EP006) public {
+    function E00010005(string memory _eventId, int256 _EP001, int256 _EP002, int256 _EP003, int256 _EP004, int256 _EP006) public {
         uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -86,11 +86,11 @@ contract TransactionContract {
         transaction.params["EP002"] = _EP002;
         transaction.params["EP003"] = _EP003;
         transaction.params["EP004"] = _EP004;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
         transaction.params["EP006"] = _EP006;
     }
 
-    function E00010006(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _EP003,uint256 _EP004, uint256 _EP006) public {
+    function E00010006(string memory _eventId, int256 _EP001, int256 _EP002, int256 _EP003,int256 _EP004, int256 _EP006) public {
        uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -102,11 +102,11 @@ contract TransactionContract {
         transaction.params["EP002"] = _EP002;
         transaction.params["EP003"] = _EP003;
         transaction.params["EP004"] = _EP004;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
         transaction.params["EP006"] = _EP006;
     }
 
-    function E00020001(string memory _eventId, uint256 _EP001, uint256 _EP002, uint256 _BTC, uint256 _USDT) public {
+    function E00020001(string memory _eventId, int256 _EP001, int256 _EP002, int256 _BTC, int256 _USDT) public {
        uint256 index = transactions.length;
         transactions.push();
         Transaction storage transaction = transactions[index];
@@ -118,7 +118,7 @@ contract TransactionContract {
         transaction.params["EP002"] = _EP002;
         transaction.params["BTC"] = _BTC;
         transaction.params["USDT"] = _USDT;
-        transaction.params["trans_time"] = block.timestamp;
+        transaction.params["trans_time"] = int256(block.timestamp);
     }
     
     function getTransaction(string memory _eventId) public view returns (string memory, string memory, address) {
@@ -151,10 +151,10 @@ contract TransactionContract {
         return transactions[index].eventId;
     }
 
-    function getTransactionParam(uint256 index, string memory paramKey) public view returns (uint256) {
+    function getTransactionParam(uint256 index, string memory paramKey) public view returns (int256) {
         return transactions[index].params[paramKey];
     }
-    function getTransactionParamByEventId(string memory _eventId, string memory _paramKey) public view returns (uint256) {
+    function getTransactionParamByEventId(string memory _eventId, string memory _paramKey) public view returns (int256) {
         Transaction storage transaction = findTransaction(_eventId);
         return transaction.params[_paramKey];
     }
